requirements.txt
streamlit
requests
pandas
plotly
python-dotenv
Pillow
google-generativeai
groq
python-dateutil